requirements.txt
streamlit
firebase-admin
bcrypt
pandas
plotly
pyrebase4
setuptools
Pillow
